Implement new CryptedFileKeyring.
CryptedFileKeyring now uses PBKDF2 to derive the key from the user's password
and a random hash. The IV is chosen randomly as well. All the stored passwords
are encrypted at once. A method to convert the old format to the new one is
provided.

The user's password is not limited to 32 characters anymore.